add tests for AsyncCallServer
